+++ src/spring/HISTORY.md
@@ -1,5 +1,8 @@
 Release History
 ===============
+1.19.3
+---
+* Add arguments `--refresh-interval` in `spring application-configuration-service create` and `spring application-configuration-service update`.
 
 1.19.2
 ---